% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods_therMizer.R
\docType{methods}
\name{getESpawning}
\alias{getESpawning}
\alias{getESpawning,therMizerParams,matrix,numeric,matrix-method}
\title{getESpawning method for the size based model}
\usage{
getESpawning(object, n, n_pp, e)

\S4method{getESpawning}{therMizerParams,matrix,numeric,matrix}(object, n, n_pp,
  e)
}
\arguments{
\item{object}{A \code{therMizerParams} object.}

\item{n}{A matrix of species abundance (species x size).}

\item{n_pp}{A vector of the background abundance by size.}

\item{e}{The energy available for reproduction and growth (optional). A
matrix of size no. species x no. size bins. If not supplied, is calculated
internally using the \code{getEReproAndGrowth()} method.}

\item{ocean_temp}{A numeric vector of the ocean temperature by realm or a 
single numeric ocean_temp value which is used for all realms.}
}
\value{
A two dimensional array (prey species x prey size)
}
\description{
Calculates the energy available by species and size for reproduction after
metabolism and movement have been accounted for.
Used by the \code{project_therMizer} method for performing simulations.

\code{getESpawning} method with \code{e} argument.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- therMizerParams(NS_species_params_gears, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project_therMizer(params, t_max = 20, effort = 0.5)
# Get the energy at a particular time step
getESpawning(params,sim@n[21,,],sim@n_pp[21,])
}
}
\seealso{
\code{\link{project_therMizer}} and \code{\link{getEReproAndGrowth}}.
}
