% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods_therMizer.R
\docType{methods}
\name{getSSA}
\alias{getSSA}
\alias{getSSA,therMizerSim-method}
\title{Calculate the SSA of species}
\usage{
getSSA(object)

\S4method{getSSA}{therMizerSim}(object)
}
\arguments{
\item{object}{An object of class \code{therMizerSim}.}
}
\value{
An array containing the SSA (time x species)
}
\description{
Calculates the spawning stock abundance (SSA) through time of the species in 
the \code{therMizerSim} class. SSA is calculated as the total abundance of all mature 
individuals.

Calculate the SSA from a \code{therMizerSim} object
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- therMizerParams(NS_species_params_gears, inter)
# With constant fishing effort for all gears for 20 time steps
sim <- project_therMizer(params, t_max = 20, effort = 0.5)
getSSA(sim)
}
}
