% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods_therMizer.R
\docType{methods}
\name{getTempEffectMetab}
\alias{getTempEffectMetab}
\alias{getTempEffectMetab,therMizerParams,numeric-method}
\alias{getTempEffectMetab,therMizerParams,matrix-method}
\alias{getTempEffectMetab,therMizerSim,missing-method}
\title{Get the total temperature effect on metabolism across all realms by time, species, and size.}
\usage{
getTempEffectMetab(object, ocean_temp, ...)

\S4method{getTempEffectMetab}{therMizerParams,numeric}(object, ocean_temp, ...)

\S4method{getTempEffectMetab}{therMizerParams,matrix}(object, ocean_temp, ...)

\S4method{getTempEffectMetab}{therMizerSim,missing}(object, ocean_temp,
  time_range = dimnames(object@ocean_temp)$time, drop = TRUE, ...)
}
\arguments{
\item{object}{A \code{therMizerParams} object or a \code{therMizerSim} object}

\item{ocean_temp}{The ocean temperature in each realm. Only needed if the object
argument is of class \code{therMizerParams}. See notes below.}

\item{...}{Other arguments passed to \code{getTempEffect} method.}

\item{time_range}{Subset the returned temperature effects by time. The time
range is either a vector of values, a vector of min and max time, or a
single value. Default is the whole time range. Only used if the
\code{object} argument is of type \code{therMizerSim}.}

\item{drop}{Only used when object is of type \code{therMizerSim}. Should
dimensions of length 1 be dropped, e.g., if your community has only one
species it might make presentation of results easier. Default is TRUE.}
}
\value{
An array. If the effort argument has a time dimension, or object is
  of class \code{therMizerSim}, the output array has three dimensions (time x 
  species x size). If the ocean_temp argument does not have a time dimension, the
  output array has two dimensions (species x size).
}
\description{
Calculates the temperature effect on metabolism across all realms by species and size at 
each time step in teh \code{ocean_temp} argument.
The total temperature effect is just the sum of the temperature effects in
each realm

\code{getTempEffectMetab} method for \code{MizerParams} object with constant temperature.

\code{getTempEffectMetab} method for \code{therMizerParams} object with time changing effort.

\code{getTempEffectMetab} method for \code{therMizerSim} object.
}
\note{
Here: temperature effect is determined from exposure x ontogenetic_migration
  x ocean_temp. 

The \code{ocean_temp} argument is only used if a \code{therMizerParams} object is
  passed in. The \code{ocean_temp} argument can be a two dimension array (time x 
  realm), a vector of length equal to the number of realms (each realm has a 
  different temperature that is constant in time), or a single numeric value 
  (each realm has the same temperature that is constant in time). The order of 
  realms in the \code{ocean_temp} argument must be the same as in the 
  \code{therMizerParams} object. 

If the object argument is of class \code{therMizerSim} then the ocean_temp slot
  of the \code{therMizerSim} object is used and the \code{ocean_temp} argument 
  is not used.
}
