% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods_therMizer.R
\docType{methods}
\name{getTempEffectRealmEncount}
\alias{getTempEffectRealmEncount}
\alias{getTempEffectRealmEncount,therMizerParams,numeric-method}
\alias{getTempEffectRealmEncount,therMizerParams,matrix-method}
\alias{getTempEffectRealmEncount,therMizerSim,missing-method}
\title{getTempEffectRealmEncount method for the size based model}
\usage{
getTempEffectRealmEncount(object, ocean_temp, ...)

\S4method{getTempEffectRealmEncount}{therMizerParams,numeric}(object,
  ocean_temp, ...)

\S4method{getTempEffectRealmEncount}{therMizerParams,matrix}(object, ocean_temp,
  ...)

\S4method{getTempEffectRealmEncount}{therMizerSim,missing}(object, ocean_temp,
  time_range = dimnames(object@ocean_temp)$time, ...)
}
\arguments{
\item{object}{A \code{therMizerParams} object or a \code{therMizerSim}
object.}

\item{ocean_temp}{The ocean temperature in each realm. Only needed if the object 
argument is of class \code{therMizerParams}. See notes below.}

\item{...}{Other arguments (currently unused)}

\item{time_range}{Subset the returned temperature effects by time. The time
range is either a vector of values, a vector of min and max time, or a 
single value. Default is the whole time range. Only used if the \code{object}
argument is of type \code{therMizerSim}.}
}
\value{
An array. If the ocean_temp argument has a time dimension, or a 
  \code{therMizerSim} is passed in, the output array has four dimensions
  (time x realm x species x size). If the ocean_temp argument does not have a 
  time dimension (i.e., it is a vector or a single numeric), the output 
  array has three dimensions (realm x species x size).
}
\description{
Calcuates the temperature effect on encounter rate by realm, species, and size at each time step
in the \code{ocean_temp} argument. Used by the \code{project_therMizer} method to 
perform simulations

\code{getTempEffectRealmEncount} method for \code{therMizerParams} object with constant temperature.

\code{getTempEffectRealmEncount} method for \code{therMizerParams} object with time-changing temp.

\code{getTempEffectRealmEncount} method for \code{therMizerSim} object.
}
\note{
Here: Temperature effect results from exposure x ontogenetic migration
  x ocean temperature. 

The \code{ocean_temp} argument is only used of a \code{therMizerParams} 
  object is passed in. The \code{ocean_temp} argument can be a two dimensional
  array (time x realm), a vector of length equal to the number of realms 
  (each realm has a different temperature that is constant in time), or a
  single numeric value (each realm has the same temperature that is constant
  in time). The order of realms in the \code{ocean_temp} argument must be the
  same as in the \code{therMizerParams} object. 

If the object argument is of class \code{therMizerSim} then the 
  ocean_temp slot of the \code{therMizerSim} object is used and the 
  \code{ocean_temp} argument is not used.
}
