% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods_therMizer.R
\docType{methods}
\name{getYield}
\alias{getYield}
\alias{getYield,therMizerSim-method}
\title{Calculate the total yield of each species}
\usage{
getYield(object, ...)

\S4method{getYield}{therMizerSim}(object, ...)
}
\arguments{
\item{object}{An object of class \code{therMizerSim}.}
}
\value{
An array containing the total yield (time x species)
}
\description{
Calculates the total yield of each species across all gears at each
simulation time step.

Calculate the total yield of each species from a \code{therMizerSim} object
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- therMizerParams(NS_species_params_gears, inter)
sim <- project_therMizer(params, effort=1, t_max=10)
y <- getYield(sim)
}
}
\seealso{
\code{\link{getYieldGear}}
}
