% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods_therMizer.R
\docType{methods}
\name{getYieldGearSize}
\alias{getYieldGearSize}
\alias{getYieldGearSize,therMizerSim-method}
\title{Calculate the total yield per gear, species, and size}
\usage{
getYieldGearSize(object, ...)

\S4method{getYieldGearSize}{therMizerSim}(object, ...)
}
\arguments{
\item{object}{An object of class \code{therMizerSim}.}
}
\value{
An array containing the total yield (time x gear x species x size)
}
\description{
Calculates the total yield per gear, species, and size at each simulation
time step.

Calculate the total yield per gear and size from a \code{therMizerSim} object
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- therMizerParams(NS_species_params_gears, inter)
# With constant fishing effort for all gears for 20 time steps
sim <- project_therMizer(params, t_max = 20, effort = 0.5)
getYieldGear(sim)
}
}
\seealso{
\code{\link{getYield}}
}
