% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity_funcs_therMizer.R
\name{knife_edge_phased}
\alias{knife_edge_phased}
\title{Size based knife-edge selectivity function that is phased in}
\usage{
knife_edge_phased(w, knife_edge_size1, knife_edge_size2)
}
\arguments{
\item{w}{The size of the individual.}

\item{knife_edge_size1}{The size at which the knife-edge is initiated.}

\item{knife_edge_size2}{The size at which selectivity is fully realized.}
}
\description{
A knife-edge selectivity function where only sizes greater or equal to
\code{knife_edge_size1} are selected.  Selectivity is then phased in through
size \code{knife_edge_size2}.  This simulates the reduced effectiveness of 
longline gear for smaller fish
}
