% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity_funcs_therMizer.R
\name{mig_window}
\alias{mig_window}
\title{Size based window selectivity function}
\usage{
mig_window(w, mig_min_size, mig_max_size)
}
\arguments{
\item{w}{The size of the individual.}

\item{mig_min_size}{The minimum window size.}

\item{mig_max_size}{The maximum window size.}
}
\description{
Essentiall a double-knife-edge selectivity function where only sizes 
within a range given by \code{mig_min_size} and \code{mig_max_size}
are selected.
}
