% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_therMizer.R
\docType{methods}
\name{plot,therMizerSim,missing-method}
\alias{plot,therMizerSim,missing-method}
\title{Summary plot for \code{therMizerSim} objects}
\usage{
\S4method{plot}{therMizerSim,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{therMizerSim}}

\item{y}{Not used}

\item{...}{For additional arguments see the documentation for
\code{\link{plotBiomass}},
\code{\link{plotFeedingLevel}},\code{\link{plotM2}},
\code{\link{plotTempEffectEncount}},
\code{\link{plotTempEffectMetab}},
and \code{\link{plotFMort}}.}
}
\value{
A viewport object
}
\description{
After running a projection, produces 6 plots in the same window: feeding
level, predation mortality, fishing mortality, and temperature effects on
metabolism and encounter rate for each species by size; and biomass of 
each species through time. This method just uses the other plotting 
methods and puts them all in one window.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- therMizerParams(NS_species_params_gears, inter)
sim <- project_therMizer(params, effort=1, t_max=20, t_save = 2)
plot(sim)
plot(sim, time_range = 10:20) # change time period for size-based plots
plot(sim, min_w = 10, max_w = 1000) # change size range for biomass plot
}
}
