% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_therMizer.R
\docType{methods}
\name{plotTempEffectMetab}
\alias{plotTempEffectMetab}
\alias{plotTempEffectMetab,therMizerSim-method}
\title{Plot total temperature effect on metabolism on each species by size}
\usage{
plotTempEffectMetab(object, ...)

\S4method{plotTempEffectMetab}{therMizerSim}(object,
  time_range = max(as.numeric(dimnames(object@n)$time)), print_it = TRUE,
  ...)
}
\arguments{
\item{object}{An object of class \code{therMizerSim}.}

\item{...}{Other arguments to pass to \code{getFMort} method}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step.}

\item{print_it}{Display the plot, or just return the ggplot2 object}
}
\value{
A ggplot2 object
}
\description{
After running a projection, plot the temperature effect on metabolism for 
each species by size. The total temperature effect is averaged over the 
specified time range (a single value for the time range can be used to 
plot a single time step).

Plot total temperature effect using a \code{therMizerSim} object.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- therMizerParams(NS_species_params_gears, inter)
sim <- project_therMizer(params, effort=1, t_max=20, t_save = 2)
plotTempEffectMetab(sim)
plotTempEffectMetab(sim, time_range = 10:20)
}
}
\seealso{
\code{\link{getTempEffect}}
}
