% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods_therMizer.R
\docType{methods}
\name{summary,therMizerParams-method}
\alias{summary,therMizerParams-method}
\title{Summarize therMizerParams object}
\usage{
\S4method{summary}{therMizerParams}(object, ...)
}
\arguments{
\item{object}{A \code{therMizerParams} object.}

\item{...}{Other arguments (currently not used).}
}
\description{
Outputs a general summary of the structure and content of the object
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- therMizerParams(NS_species_params_gears, inter)
summary(params)
}
}
