% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods_therMizer.R
\docType{methods}
\name{summary,therMizerSim-method}
\alias{summary,therMizerSim-method}
\title{Summarize therMizerSim object}
\usage{
\S4method{summary}{therMizerSim}(object, ...)
}
\arguments{
\item{object}{A \code{therMizerSim} object.}

\item{...}{Other arguments (currently not used).}
}
\description{
Outputs a general summary of the structure and content of the object
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- therMizerParams(NS_species_params_gears, inter)
sim <- project_therMizer(params, effort=1, t_max=5)
summary(sim)
}
}
