% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/therMizerSim-class.R
\docType{class}
\name{therMizerSim-class}
\alias{therMizerSim-class}
\title{therMizerSim}
\description{
A class that holds the results of projecting a \linkS4class{therMizerParams}
object through time.
}
\details{
\code{therMizerSim} objects are created by using the \code{\link{project_therMizer}} method
on an object of type \code{therMizerParams}.

There are several plotting methods available to explore the contents of a
\code{therMizerSim} object. See the package vignette for more details.
}
\section{Slots}{

\describe{
\item{\code{params}}{An object of type \linkS4class{therMizerParams}.}

\item{\code{n}}{Array that stores the projected community population abundances by
time, species and size}

\item{\code{effort}}{Array that stores the fishing effort through time by time and
gear}

\item{\code{ocean_temp}}{Array that stores the ocean temperature through time by
time and realm}

\item{\code{n_pp}}{Array that stores the projected background population by time and
size}
}}

\seealso{
\code{\link{project_therMizer}} \code{\link{therMizerParams}}
}
