% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/therMizerSim-class.R
\docType{methods}
\name{therMizerSim}
\alias{therMizerSim}
\alias{therMizerSim,therMizerParams-method}
\title{Constructor for the \code{therMizerSim} class}
\usage{
therMizerSim(object, ...)

\S4method{therMizerSim}{therMizerParams}(object, t_dimnames = NA,
  t_max = 100, t_save = 1)
}
\arguments{
\item{object}{a \linkS4class{therMizerParams} object}

\item{...}{Other arguments (currently not used).}

\item{t_dimnames}{Numeric vector that is used for the time dimensions of the
slots. Default = NA.}

\item{t_max}{The maximum time step of the simulation. Only used if t_dimnames
= NA. Default value = 100.}

\item{t_save}{How often should the results of the simulation be stored. Only
used if t_dimnames = NA. Default value = 1.}
}
\value{
An object of type \linkS4class{therMizerSim}
}
\description{
A constructor for the \code{therMizerSim} class. This is used by the
\code{project_therMizer} method to create \code{therMizerSim} objects of the right
dimensions. It is not necessary for users to use this constructor.

therMizerSim constructor taking only a \code{therMizerParams} object.
}
\examples{
\dontrun{
data(NS_species_params_gears)
data(inter)
params <- therMizerParams(NS_species_params_gears, inter)
sim <- project_therMizer(params)
}
}
\seealso{
\code{\link{project_therMizer}} \linkS4class{therMizerParams}
  \linkS4class{therMizerSim}
}
